Implements support for list input
It is not inconcievable to want to cut more then one directory in one
go, particulary in the case of downstreams.

I'm not 100% that we should be avoiding just targeting the whole project
in cases like ss13, but a list input is pretty easy to implement so may
as well.

I am considering doing the same thing to the templates string but that
is mildly more complex.