Implements diagonal edge smoothing, bumps bitflags version
Adds an enum that controls the type of output we produce, either
cardinals, all, or all + inner/outer edges to do wall diagonals.

We do this by adding a new type of adjacency that we can optionally
attempt to fill, and then a more formal definition of corner types
pulled from afformentioned enum. This could all just be done with
prefabs but I prefer having formal support for it.